Created package with multiple entry points
