+++ .gitignore
@@ -1,3 +1,6 @@
+# egg info
+simview.egg-info
+
 # Compiled Object files
 *.slo
 *.lo
