--- package/__init__.py
@@ -1,11 +1,0 @@
-""" SimView
-An AR/VR molecular simulation viewer.
-"""
-
-# need to import md to have library available.
-import hoomd.md
-# import make reverse for testing purposes
-from .smiles_sim import *
-import .smarter_broker
-import .record_client
-
