+++ simview/__init__.py
@@ -1,0 +1,6 @@
+""" SimView
+An AR/VR molecular simulation viewer.
+"""
+
+from . import smarter_broker, smiles_sim
+from .simulation import SimulationChannel
